// packages/substore-overlay/src/types/clash_meta_config.ts
export type ClashMetaConfig = NoName &
  NoName2 &
  Iptables &
  TLS &
  NoName13 &
  NoName29 &
  NoName284 &
  NoName286 &
  Listeners &
  NoName386 &
  NoName388 &
  NoName409 &
  Hosts &
  NTP &
  DNS6 &
  TUN3 &
  TuicServer &
  NoName453 &
  NoName455 &
  NoName460;
// ...
export type NoName13 = Inbound &
  Controller & {
    "unified-delay"?: NoName21;
    mode?: NoName22;
    "log-level"?: NoName23;
    ipv6?: IPv6;
    "interface-name"?: NoName24;
    "routing-mark"?: NoName25;
    "geox-url"?: GeoxUrl;
    "geo-auto-update"?: Geo;
    "geo-update-interval"?: Geo1;
    "geodata-mode"?: GeoIP;
    "geodata-loader"?: Geo2;
    "geosite-matcher"?: Geosite1;
    "tcp-concurrent"?: TCP;
    "find-process-mode"?: NoName26;
    "global-client-fingerprint"?: NoName27;
    "global-ua"?: UA;
    "etag-support"?: ETag;
    "keep-alive-idle"?: S;
    "keep-alive-interval"?: S1;
    "disable-keep-alive"?: NoName28;
    [k: string]: unknown;
  };
// ...
export type NoName21 = boolean;
/**
 * 运行模式
 */
export type NoName22 = "global" | "rule" | "direct";
/**
 * 日志级别
 */
export type NoName23 = "debug" | "info" | "warning" | "error" | "silent";
/**
 * 是否允许内核接受IPv6流量
 */
export type IPv6 = boolean;
/**
 * 流量出站接口
 */
export type NoName24 = string;
/**
 * 为Linux下的出站连接提供默认流量标记
 */
export type NoName25 = number;
/**
 * geoip文件下载地址
 */
export type Geoip = string;
/**
 * mmdb文件下载地址
 */
export type Mmdb = string;
/**
 * asn文件下载地址
 */
export type Asn = string;
/**
 * geosite文件下载地址
 */
export type Geosite = string;
/**
 * Geo自动更新
 */
export type Geo = boolean;
/**
 * 单位:h
 */
export type Geo1 = number;
/**
 * 更改geoip使用文件(true: dat, false: mmdb)
 */
export type GeoIP = boolean;
/**
 * Geo文件加载器
 */
export type Geo2 = "standard" | "memconservative" | "memc";
/**
 * Geosite匹配器
 */
export type Geosite1 = "mph" | "hybrid" | "succinct";
/**
 * TCP并发
 */
export type TCP = boolean;
/**
 * 进程匹配模式
 */
export type NoName26 = "always" | "strict" | "off";
/**
 * 支持协议有VLESS,Vmess,trojan
 */
export type NoName27 =
  | "chrome"
  | "chrome_psk"
  | "chrome_psk_shuffle"
  | "chrome_padding_psk_shuffle"
  | "chrome_pq"
  | "chrome_pq_psk"
  | "firefox"
  | "safari"
  | "ios"
  | "android"
  | "edge"
  | "360"
  | "qq"
  | "random"
  | "randomized";
/**
 * 下载外部资源(proxy-provider、rule-provider、Geo等)时使用
 */
export type UA = string;
/**
 * ETag是一种用于缓存和条件请求的机制
 */
export type ETag = boolean;
/**
 * 连接空闲后多久发送探测包
 */
export type S = number;
/**
 * 发送探测包的时间间隔
 */
export type S1 = number;
/**
 * 禁用保持活跃探测
 */
export type NoName28 = boolean;
// ...
export type NoName30 = (
  | Direct
  | Dns
  | Shadowsocks1
  | Shadowsocksr
  | Snell
  | Socks5
  | Http
  | Vmess1
  | Vless
  | Trojan
  | Hysteria
  | Hysteria2
  | Wireguard
  | Tuic
  | Ssh
  | Mieru
  | Anytls
  | Sudoku
  | Masque
)[];
// ...
export type NoName287 = (Selector | UrlTest | Fallback | LoadBalance | Relay)[];
// ...
export type Listeners1 = (
  | Socks
  | Http2
  | Tproxy
  | Redir
  | Mixed
  | Tunnel
  | TUN
  | Shadowsocks4
  | Vmess3
  | Vless2
  | Trojan2
  | Hysteria22
  | Tuic2
  | Anytls2
  | Mieru2
  | Sudoku2
)[];
export type NoName410 = ((
  | NoName390
  | NoName391
  | NoName392
  | NoName393
  | NoName394
  | Geosite2
  | IP8
  | IP9
  | IPASN
  | IPASN1
  | IP10
  | IP11
  | IP12
  | IP13
  | IP14
  | NoName395
  | NoName396
  | NoName397
  | DSCPTproxyUdp
  | Android3
  | NoName398
  | Android4
  | NoName399
  | Android5
  | NoName400
  | TcpUdp
  | LinuxUSERID
  | NoName401
  | NoName402
  | NoName403
  | NoName404
  | RuleProviders
  | NoName405
  | NoName406
  | NoName407
  | NoName408
) &
  string)[];
/**
 * NTP(Network Time Protocol),用于同步计算机系统时钟的网络协议
 */
export type NTP2 = boolean;
/**
 * NTP服务器地址
 */
export type NTP3 = string;
/**
 * NTP服务器端口
 */
export type NTP4 = number | string;
/**
 * 单位:m
 */
export type NoName411 = number;
/**
 * 取值为proxy-groups,proxy的name
 */
export type ProxyDialerProxy = string;
/**
 * 需要root/管理员权限
 */
export type NoName412 = boolean;
/**
 * DNS配置
 */
export type Dns3 = {
  [k: string]: unknown;
} & {
  enable?: DNS7;
  "prefer-h3"?: DoHHTTP3;
  ipv6?: IPV6;
  "ipv6-timeout"?: IPV61;
  "use-hosts"?: Hosts2;
  "use-system-hosts"?: Hosts3;
  "respect-rules"?: DNS8;
  nameserver?: NoName413;
  fallback?: NoName414;
  "fallback-filter"?: NoName415;
  listen?: DNS11;
  "enhanced-mode"?: NoName416;
  "fake-ip-range"?: FakeIpIpv4;
  "fake-ip-range6"?: FakeIpIpv6;
  "fake-ip-filter"?: FakeIp1;
  "fake-ip-filter-mode"?: FakeIp2;
  "fake-ip-ttl"?: FakeipTTL;
  "default-nameserver"?: NoName417;
  "cache-algorithm"?: NoName418;
  "cache-max-size"?: NoName419;
  "nameserver-policy"?: DNS12;
  "proxy-server-nameserver"?: DNS13;
  "proxy-server-nameserver-policy"?: DNS14;
  "direct-nameserver"?: DirectDNS;
  "direct-nameserver-follow-policy"?: DNS15;
  [k: string]: unknown;
};
// ...
export type NoName454 = unknown[];
/**
 * 可为tcp/udp
 */
export type NoName456 = ("tcp" | "udp")[];
/**
 * 本地监听地址
 */
export type NoName457 = {
  [k: string]: unknown;
};
/**
 * 转发的目标地址
 */
export type NoName458 = string;
/**
 * 经由某个proxy/proxy-group发送流量(可选)
 */
export type NoName459 = string;
/**
 * 隧道配置
 */
export type Tunnels = (
  | {
      [k: string]: unknown;
    }
  | {
      network?: NoName456;
      address?: NoName457;
      target?: NoName458;
      proxy?: NoName459;
      [k: string]: unknown;
    }
)[];
// ...
export interface NoName {
  experimental?: NoName1;
  [k: string]: unknown;
}
export interface NoName1 {
  "quic-go-disable-gso"?: QuicGoGSO;
  "quic-go-disable-ecn"?: QuicGoECN;
  "dialer-ip4p-convert"?: IP4P;
  [k: string]: unknown;
}
export interface NoName2 {
  profile?: NoName3;
  [k: string]: unknown;
}
export interface NoName3 {
  "store-selected"?: NoName4;
  "store-fake-ip"?: FakeIp;
  [k: string]: unknown;
}
export interface Iptables {
  iptables?: Iptables1;
  [k: string]: unknown;
}
export interface Iptables1 {
  enable?: NoName5;
  "inbound-interface"?: NoName6;
  bypass?: IptablesIP;
  "dns-redirect"?: DNS;
  [k: string]: unknown;
}
export interface TLS {
  tls?: TLS1;
  [k: string]: unknown;
}
export interface TLS1 {
  certificate?: NoName8;
  "private-key"?: NoName9;
  "client-auth-type"?: NoName10;
  "client-auth-cert"?: NoName11;
  "ech-key"?: ECH;
  "custom-certifactes"?: NoName12;
  [k: string]: unknown;
}
export interface Inbound {
  port?: HTTPS;
  "socks-port"?: SOCKS;
  "redir-port"?: NoName14;
  "tproxy-port"?: NoName15;
  "mixed-port"?: HTTPSSOCKS;
  "ss-config"?: Shadowsocks;
  "vmess-config"?: Vmess;
  "allow-lan"?: NoName16;
  "skip-auth-prefixes"?: IP;
  "lan-allowed-ips"?: IP1;
  "lan-disallowed-ips"?: IP2;
  "bind-address"?: IP3 | IP4;
  "inbound-tfo"?: TFOTCPFastOpen;
  "inbound-mptcp"?: MPTCPMultiPathTCP;
  [k: string]: unknown;
}
export interface Controller {
  "external-controller"?: HTTP;
  "external-controller-tls"?: HTTPS1;
  "external-controller-pipe"?: WindowsNamedpipe;
  "external-controller-unix"?: UnixSocket;
  "external-controller-cors"?: CORS;
  "external-ui"?: Ui;
  "external-ui-url"?: NoName19;
  "external-ui-name"?: NoName20;
  "external-doh-server"?: DOH;
  secret?: RESTfulAPI;
  [k: string]: unknown;
}
/**
 * 跨域资源共享(CORS)配置
 */
export interface CORS {
  "allow-origins"?: NoName17;
  "allow-private-network"?: NoName18;
  [k: string]: unknown;
}
export interface GeoxUrl {
  geoip?: Geoip;
  mmdb?: Mmdb;
  asn?: Asn;
  geosite?: Geosite;
  [k: string]: unknown;
}
export interface NoName29 {
  proxies?: NoName30;
  [k: string]: unknown;
}
// ...
export interface NoName284 {
  "proxy-providers"?: NoName285;
  [k: string]: unknown;
}
export interface NoName285 {
  [k: string]: {
    [k: string]: unknown;
  };
}
export interface NoName286 {
  "proxy-groups"?: NoName287;
  [k: string]: unknown;
}
// ...
export interface Listeners {
  listeners?: Listeners1;
  [k: string]: unknown;
}
// ...
export interface NoName386 {
  "rule-providers"?: NoName387;
  [k: string]: unknown;
}
export interface NoName387 {
  [k: string]: {
    [k: string]: unknown;
  };
}
export interface NoName388 {
  "sub-rules"?: NoName389;
  [k: string]: unknown;
}
export interface NoName389 {
  [k: string]: ((
    | NoName390
    | NoName391
    | NoName392
    | NoName393
    | NoName394
    | Geosite2
    | IP8
    | IP9
    | IPASN
    | IPASN1
    | IP10
    | IP11
    | IP12
    | IP13
    | IP14
    | NoName395
    | NoName396
    | NoName397
    | DSCPTproxyUdp
    | Android3
    | NoName398
    | Android4
    | NoName399
    | Android5
    | NoName400
    | TcpUdp
    | LinuxUSERID
    | NoName401
    | NoName402
    | NoName403
    | NoName404
    | RuleProviders
    | NoName405
    | NoName406
    | NoName407
    | NoName408
  ) &
    string)[];
}
// ...
export interface NoName409 {
  rules?: NoName410;
  [k: string]: unknown;
}
export interface Hosts {
  hosts?: Hosts1;
  [k: string]: unknown;
}
/**
 * key: value格式(域名支持通配)
 */
export interface Hosts1 {
  [k: string]: unknown;
}
export interface NTP {
  ntp?: NTP1;
  [k: string]: unknown;
}
export interface NTP1 {
  enable?: NTP2;
  server?: NTP3;
  port?: NTP4;
  interval?: NoName411;
  "dialer-proxy"?: ProxyDialerProxy;
  "write-to-system"?: NoName412;
  [k: string]: unknown;
}
export interface DNS6 {
  dns?: Dns3;
  [k: string]: unknown;
}
// ...
export interface TUN3 {
  tun?: Tun1;
  [k: string]: unknown;
}
export interface Tun1 {
  enable?: TUN4;
  device?: TUN5;
  stack?: TUN6;
  "dns-hijack"?: DNS16;
  "auto-route"?: NoName420;
  "auto-detect-interface"?: NoName421;
  mtu?: NoName422;
  gso?: NoName423;
  "gso-max-size"?: NoName424;
  "inet4-address"?: TUNIPv41;
  "inet6-address"?: TUNIPv61;
  "iproute2-table-index"?: NoName425;
  "iproute2-rule-index"?: NoName426;
  "auto-redirect"?: NoName427;
  "auto-redirect-input-mark"?: NoName428;
  "auto-redirect-output-mark"?: NoName429;
  "loopback-address"?: NoName430;
  "strict-route"?: NoName431;
  "route-address"?: AutoRoute2;
  "route-address-set"?: IPCIDR2;
  "route-exclude-address"?: AutoRoute3;
  "route-exclude-address-set"?: IPCIDR3;
  "include-interface"?: NoName432;
  "exclude-interface"?: NoName433;
  "include-uid"?: Linux2;
  "include-uid-range"?: LinuxTun2;
  "exclude-uid"?: Linux3;
  "exclude-uid-range"?: LinuxTun3;
  "exclude-src-port"?: NoName434;
  "exclude-src-port-range"?: NoName435;
  "exclude-dst-port"?: NoName436;
  "exclude-dst-port-range"?: NoName437;
  "include-android-user"?: Android6;
  "include-package"?: Android7;
  "exclude-package"?: Android8;
  "endpoint-independent-nat"?: NAT1;
  "udp-timeout"?: UDPNAT1;
  "disable-icmp-forwarding"?: ICMP1;
  "file-descriptor"?: NoName438;
  "inet4-route-address"?: AutoRouteIPV42;
  "inet6-route-address"?: AutoRouteIPV62;
  "inet4-route-exclude-address"?: AutoRouteIPV43;
  "inet6-route-exclude-address"?: AutoRouteIPV63;
  recvmsgx?: NoName439;
  sendmsgx?: NoName440;
  [k: string]: unknown;
}
export interface TuicServer {
  "tuic-server"?: TuicServer1;
  [k: string]: unknown;
}
export interface TuicServer1 {
  enable: TuicServer2;
  listen: NoName441;
  token?: Token;
  users?: KeyValue4;
  certificate?: NoName442;
  "private-key"?: NoName443;
  "congestion-controller"?: NoName444;
  "max-idle-time"?: NoName445;
  "authentication-timeout"?: NoName446;
  alpn?: ALPN9;
  "max-udp-relay-packet-size"?: UDP19;
  "max-datagram-frame-size"?: NoName447;
  cwnd?: NoName448;
  "mux-option"?: NoName449;
  [k: string]: unknown;
}
/**
 * uuid:password
 */
export interface KeyValue4 {
  [k: string]: string;
}
export interface NoName449 {
  padding?: NoName450;
  brutal?: Brutal4;
  [k: string]: unknown;
}
export interface Brutal4 {
  enabled?: Brutal5;
  up?: NoName451;
  down?: NoName452;
  [k: string]: unknown;
}
export interface NoName453 {
  authentication?: NoName454;
  [k: string]: unknown;
}
export interface NoName455 {
  tunnels?: Tunnels;
  [k: string]: unknown;
}
export interface NoName460 {
  sniffer?: Sniffer;
  [k: string]: unknown;
}
/**
 * 域名嗅探配置
 */
export interface Sniffer {
  enable?: NoName461;
  "override-destination"?: NoName462;
  "force-domain"?: NoName463;
  "skip-domain"?: NoName466;
  "skip-src-address"?: IP15;
  "skip-dst-address"?: IP16;
  sniffing?: NoName467;
  "port-whitelist"?: NoName468;
  "force-dns-mapping"?: RedirHost;
  "parse-pure-ip"?: NoName469;
  sniff?: NoName470;
  [k: string]: unknown;
}